requirements.txt
# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Azure OpenAI
openai>=1.0.0

# Image processing
opencv-python>=4.8.0
pillow>=10.0.0
google-cloud-vision>=3.7.2

# HTTP requests
requests>=2.31.0

# YAML parsing
pyyaml>=6.0.1

# CORS
python-multipart>=0.0.6

# Azure Blob Storage
azure-storage-blob>=12.19.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
